Change to src-layout, use pyproject.toml dependencies
