+++ src/ephys_link/platforms/new_scale_manipulator.py
@@ -14,8 +14,8 @@ import socketio
 # noinspection PyUnresolvedReferences
 from NstMotorCtrl import NstCtrlAxis
 
-import ephys_link.common as com
-from ephys_link.platform_manipulator import PlatformManipulator
+import src.ephys_link.common as com
+from src.ephys_link.platform_manipulator import PlatformManipulator
 
 # Constants
 ACCELERATION_MULTIPLIER = 5
