+++ src/ephys_link/resources/CP210xManufacturing.dll
