+++ tests/drive_to_depth_test.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock
 # noinspection PyPackageRequirements
 import socketio
 
-from ephys_link.common import DriveToDepthOutputData
+from src.ephys_link import DriveToDepthOutputData
 
 
 # noinspection DuplicatedCode
