+++ tests/move_test.py
@@ -5,7 +5,7 @@ from unittest.mock import Mock
 # noinspection PyPackageRequirements
 import socketio
 
-from ephys_link.common import PositionalOutputData
+from src.ephys_link import PositionalOutputData
 
 
 # noinspection DuplicatedCode
